+++ vlm/__main__.py
@@ -1,24 +1,7 @@
 from aiohttp import web
 import logging
-import traceback
 
-logger = logging.getLogger(__name__)
-
-
-def _handle_exception(e, request):
-    logger.error(f'{request.headers.get("From")} "{e}"')
-    raise e
-
-
-@web.middleware
-async def error_middleware(request, handler):
-    try:
-        return await handler(request)
-    except web.HTTPError as e:
-        _handle_exception(e, request)
-    except Exception as e:
-        error_reason = f'{e}: {traceback.format_exc()}'
-        _handle_exception(web.HTTPInternalServerError(reason=error_reason), request)
+from vlm.web_app import init_web_app
 
 
 async def status(request: web.Request) -> web.Response:
@@ -27,10 +10,7 @@ async def status(request: web.Request) -> web.Response:
 
 def run():
     logging.basicConfig(level=logging.INFO)
-    app = web.Application(middlewares=[error_middleware], client_max_size=(1024 ** 2) * 10)
-    app.add_routes([
-        web.get('/status', status),
-    ])
+    app = init_web_app()
     web.run_app(
         app,
         host='0.0.0.0',  # nosec
