+++ vlm/test_vlm.py
@@ -1,0 +1,15 @@
+from aiohttp.test_utils import AioHTTPTestCase
+
+from vlm.web_app import init_web_app
+
+
+class VlmTestCase(AioHTTPTestCase):
+
+    async def get_application(self):
+        return await init_web_app()
+
+    async def test_status(self):
+        async with self.client.request('GET', '/status') as resp:
+            self.assertEqual(resp.status, 200)
+            resp_json = await resp.json()
+        self.assertDictEqual(resp_json, {'success': True})
